Updated createDIr to semi-final state. Current state requires a merge from the parseattrs branch. Fixed glfs_mkdir in GLFS.java and included a test case for it in GLFSTest.java. Changed the test for openDir to attempt to open the directory created using glfs_mkdir rather than root. Could not test changes for unknown reasons (testInit failing).

+++ libgfapi-jni/libgfapi-jni/src/main/java/com/peircean/libgfapi_jni/internal/GLFS.java
@@ -170,9 +170,9 @@ public class GLFS {
             int mode
     );
 
-    //    glfs_fd_t* glfs_mkdir(glfs_t *fs, const char *path, int mode);
+    //    int glfs_mkdir(glfs_t *fs, const char *path, int mode);
     @JniMethod(cast = "glfs_fd_t *")
-    public static final native long glfs_mkdir(
+    public static final native int glfs_mkdir(
             @JniArg(cast = "glfs_t *") long fs,
             @JniArg(cast = "const char *") String path,
             int mode
